requirements.txt
geopandas==0.10.2
rasterio==1.2.10
fiona==1.8.21
tf-models-official==2.8.0
protobuf==3.19.4
tensorflow-addons==0.16.1
tensorflow-datasets==4.5.2
tensorflow-hub==0.12.0
tensorflow-metadata==1.7.0
typing-extensions==4.1.1
typeguard==2.13.3
tqdm
pyyaml 